% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{suma}
\alias{suma}
\title{suma}
\usage{
suma(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{Another number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(2, 3)
}
